Update add.h
int -> float

+++ add.h
@@ -1,4 +1,4 @@
-float dodawanie(int a, int b){
+float dodawanie(float a, float b){
 float suma;
 suma=a+b;
 return suma;
